type=appid-toolbox | introduction of new type | integrate appid migration run directly by using pan-os-php alias
