+++ appid-toolbox/old/report-generator.php
@@ -20,9 +20,14 @@
  */
 
 
-set_include_path(dirname(__FILE__) . '/../' . PATH_SEPARATOR . get_include_path());
-require_once dirname(__FILE__) . "/../lib/pan_php_framework.php";
-require_once dirname(__FILE__) . "/../utils/lib/UTIL.php";
+require_once dirname(__FILE__) . "/../lib/common.php";
 
-$argv[] = "phase=rule-activation";
-PH::UTILdeprecated("appid-toolbox", $argv, $argc, __FILE__);
+require_once dirname(__FILE__) . "/../lib/old/lib_2_report_generator.php";
+
+PH::print_stdout( "\n************* START OF SCRIPT ".basename(__FILE__)." ************" );
+PH::print_stdout( " - PAN-OS-PHP version: ".PH::frameworkVersion() . " [".PH::frameworkInstalledOS()."]" );
+PH::processCliArgs();
+
+logAnalysis_Phase2();
+
+PH::print_stdout( "\n************* END OF SCRIPT ".basename(__FILE__)." ************" );
