+++ appid-toolbox/rule-marker.php
@@ -19,168 +19,13 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
 
-require_once dirname(__FILE__) . "/lib/common.php";
 
-PH::print_stdout( "************* START OF SCRIPT ".basename(__FILE__)." ************" );
-PH::print_stdout( " - PAN-OS-PHP version: ".PH::frameworkVersion() . " [".PH::frameworkInstalledOS()."]" );
-PH::processCliArgs();
+set_include_path(dirname(__FILE__) . '/../' . PATH_SEPARATOR . get_include_path());
+require_once dirname(__FILE__) . "/../lib/pan_php_framework.php";
+require_once dirname(__FILE__) . "/../utils/lib/UTIL.php";
 
-ruleMarker_PhaseX();
 
-PH::print_stdout("\n************* END OF SCRIPT ".basename(__FILE__)." ************\n" );
-
-#####################################
-
-function display_usage_and_exit()
-{
-    PH::print_stdout(PH::boldText("USAGE: ")."php ".basename(__FILE__)." in=api://xxxx location=deviceGroup2 [OPTIONAL ARGS]" );
-
-
-    PH::print_stdout("Listing optional arguments:");
-    PH::print_stdout();
-
-
-
-
-
-    exit(1);
-}
-
-
-function ruleMarker_PhaseX()
-{
-    if( isset(PH::$args['help']) )
-        display_usage_and_exit();
-
-
-
-
-
-
-
-
-
-
-
-    $debugAPI = FALSE;
-
-
-
-
-    $return = shared_location();
-    $configInput = $return['configInput'];
-    $location = $return['location'];
-
-
-
-
-
-
-
-
-
-
-
-    $return = getConfig($configInput, $debugAPI);
-    $xmlDoc = $return['xmlDoc'];
-    $configOutput = $return['configOutput'];
-    $inputConnector = $return['inputConnector'];
-
-
-
-
-    $return = determineConfig($xmlDoc, $configInput, $inputConnector, $location);
-    $subSystem = $return['subSystem'];
-    $pan = $return['pan'];
-
-    PH::print_stdout(" - Found DG/Vsys '$location'");
-    PH::print_stdout(" - Looking/creating for necessary Tags to mark rules");
-    TH::createTags($pan, $configInput['type']);
-
-    //
-    // REAL JOB STARTS HERE
-    //
-    main($subSystem, $configInput, $pan, $inputConnector, $configOutput );
-}
-
-function main( $subSystem, $configInput, $pan, $inputConnector, $configOutput )
-{
-    $rules = $subSystem->securityRules->rules('!(action is.negative) and (app is.any) and !(rule is.disabled) and !(tag has appid#ignore)');
-    PH::print_stdout(" - Total number of rules: {$subSystem->securityRules->count()} vs " . count($rules) . " potentially taggable");
-
-    $ridTagLibrary = new RuleIDTagLibrary();
-    $ridTagLibrary->readFromRuleArray($subSystem->securityRules->rules());
-
-
-    PH::print_stdout("\n\n*** BEGIN TAGGING RULES ***");
-
-    $xmlPreRules = '';
-    $xmlPostRules = '';
-
-    $markedRules = 0;
-    $alreadyMarked = 0;
-
-    foreach( $rules as $rule )
-    {
-        PH::print_stdout(" - rule '{$rule->name()}'");
-
-        if( $ridTagLibrary->ruleIsTagged($rule) )
-        {
-            PH::print_stdout(" SKIPPED : already tagged");
-            $alreadyMarked++;
-            continue;
-        }
-
-        $markedRules++;
-
-
-        $newTagName = $ridTagLibrary->findAvailableTagName('appRID#');
-        PH::print_stdout();
-        PH::print_stdout("    * creating Virtual TAG '$newTagName' ... ");
-
-        PH::print_stdout("    * applying tag to rule description... ");
-
-        $newDescription = $rule->description() . ' ' . $newTagName;
-        if( strlen($newDescription) > 253 )
-            derr('description is too long, please review and edit');
-        $ridTagLibrary->addRuleToTag($rule, $newTagName);
-        $rule->setDescription($newDescription);
-
-        if( $rule->isPostRule() )
-            $xmlPostRules .= "<entry name=\"{$rule->name()}\"><description>" . htmlspecialchars($rule->description()) . "</description></entry>";
-        else
-            $xmlPreRules .= "<entry name=\"{$rule->name()}\"><description>" . htmlspecialchars($rule->description()) . "</description></entry>";
-    }
-
-    PH::print_stdout("\n\nNumber of rules marked: {$markedRules}    (vs already marked: {$alreadyMarked}");
-
-    if( $markedRules < 1 )
-        PH::print_stdout("\n\n No change to push as not rule is set to be marked");
-    else
-    {
-        if( $configInput['type'] == 'api' )
-            PH::print_stdout("\n\n**** Pushing all changes at once through API... ");
-
-
-        if( $pan->isPanorama() )
-            $xml = "<pre-rulebase><security><rules>{$xmlPreRules}</rules></security></pre-rulebase><post-rulebase><security><rules>{$xmlPostRules}</rules></security></post-rulebase>";
-        else
-            $xml = "<rulebase><security><rules>{$xmlPreRules}</rules></security></rulebase>";
-
-        if( $configInput['type'] == 'api' )
-            $inputConnector->sendSetRequest(DH::elementToPanXPath($subSystem->xmlroot), $xml);
-        else
-            // save our work !!!
-            if( $configOutput !== null )
-            {
-                if( $configOutput != '/dev/null' )
-                {
-                    $pan->save_to_file($configOutput);
-                }
-            }
-    }
-
-    PH::print_stdout("\n");
-}
+$argv[] = "phase=rule-marker";
+PH::UTILdeprecated("appid-toolbox", $argv, $argc, __FILE__);
