+++ utils/lib/UTIL.php
@@ -87,6 +87,9 @@ require_once dirname(__FILE__)."/../../migration/parser/lib/CONVERTER.php";
 require_once dirname(__FILE__)."/../../migration/parser/lib/PARSER.php";
 require_once dirname(__FILE__)."/../../migration/parser/lib/SHAREDNEW.php";
 
+require_once dirname(__FILE__)."/APPIDTOOLBOX.php";
+
+
 class UTIL
 {
     public $argv = null;
