+++ utils/lib/util_action_filter.json
@@ -937,6 +937,11 @@
         "action": [],
         "filter": []
     },
+    "appid-toolbox": {
+        "name": "appid-toolbox",
+        "action": [],
+        "filter": []
+    },
     "application": {
         "name": "application",
         "action": {
@@ -1795,6 +1800,14 @@
                     "has.vsys": {
                         "Function": {},
                         "arg": true,
+                        "ci": {
+                            "fString": "(%PROP% grp)",
+                            "input": "input\/panorama-8.0.xml"
+                        }
+                    },
+                    "with-no-serial": {
+                        "Function": {},
+                        "arg": false,
                         "ci": {
                             "fString": "(%PROP% grp)",
                             "input": "input\/panorama-8.0.xml"
